refactor: Change parameter type deleteUser method to long

+++ src/main/java/com/mycompany/SkySong/authentication/service/impl/DeleteServiceImpl.java
@@ -8,7 +8,7 @@ public class DeleteServiceImpl implements DeleteService {
     @Autowired
     private UserDAO userDAO;
     @Override
-    public String deleteUser(int userId) {
+    public String deleteUser(long userId) {
         return null;
     }
 }
